requirements-assch.txt
# A.S.S.C.H. Assembly Dependencies
# Core framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0

# Existing SalesGPT dependencies
langchain==0.1.0
openai==1.7.0
chromadb>=0.4.18
tiktoken>=0.5.2
litellm>=1.10.2
langchain-openai==0.0.2

# Optional: For production deployment
# redis>=5.0.0  # For scalable session storage
# sqlalchemy>=2.0.0  # For database persistence
# celery>=5.3.0  # For async task processing
